// web/lib/api/generated/model/shopContactChannelEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `CALL` - Call
 * * `WHATSAPP` - WhatsApp
 * * `VISIT` - Visit
 * * `OTHER` - Other
 */
export type ShopContactChannelEnum = typeof ShopContactChannelEnum[keyof typeof ShopContactChannelEnum];


export const ShopContactChannelEnum = {
  CALL: 'CALL',
  WHATSAPP: 'WHATSAPP',
  VISIT: 'VISIT',
  OTHER: 'OTHER',
} as const;