// web/lib/api/generated/model/shopContactOutcomeEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `REACHED` - Spoke to them
 * * `NO_ANSWER` - No answer
 * * `WILL_ORDER` - Will order
 * * `NOT_INTERESTED` - Not interested
 * * `OTHER` - Other
 */
export type ShopContactOutcomeEnum = typeof ShopContactOutcomeEnum[keyof typeof ShopContactOutcomeEnum];


export const ShopContactOutcomeEnum = {
  REACHED: 'REACHED',
  NO_ANSWER: 'NO_ANSWER',
  WILL_ORDER: 'WILL_ORDER',
  NOT_INTERESTED: 'NOT_INTERESTED',
  OTHER: 'OTHER',
} as const;